Changed ProductWeight to float

+++ backend/backend/models.py
@@ -9,108 +9,108 @@ from django.db import models
 
 
 class Brand(models.Model):
-    BrandID = models.AutoField(db_column='BrandID', primary_key=True)  # Field Name made lowercase.
-    BrandName = models.CharField(db_column='BrandName', max_length=45)  # Field Name made lowercase.
+	BrandID = models.AutoField(db_column='BrandID', primary_key=True)  # Field Name made lowercase.
+	BrandName = models.CharField(db_column='BrandName', max_length=45)  # Field Name made lowercase.
 
-    class Meta:
-        db_table = 'Brand'
+	class Meta:
+		db_table = 'Brand'
 
 class Category(models.Model):
-    CategoryID = models.AutoField(db_column='CategoryID', primary_key=True)  # Field Name made lowercase.
-    CategoryName = models.CharField(db_column='CategoryName', max_length=45)  # Field Name made lowercase.
+	CategoryID = models.AutoField(db_column='CategoryID', primary_key=True)  # Field Name made lowercase.
+	CategoryName = models.CharField(db_column='CategoryName', max_length=45)  # Field Name made lowercase.
 
-    class Meta:
-        db_table = 'Category'
+	class Meta:
+		db_table = 'Category'
 
 class PastOrder(models.Model):
-    OrderID = models.AutoField(db_column='OrderID', primary_key=True)  # Field Name made lowercase.
-    UserID = models.ForeignKey('User', models.DO_NOTHING, db_column='UserID')  # Field Name made lowercase.
-    OrderDelivered = models.BooleanField(db_column='OrderDelivered', default=False)  # Field Name made lowercase.
-    OrderTimestamp = models.DateTimeField(auto_now_add=True)
-    class Meta:
-        db_table = 'PastOrder'
+	OrderID = models.AutoField(db_column='OrderID', primary_key=True)  # Field Name made lowercase.
+	UserID = models.ForeignKey('User', models.DO_NOTHING, db_column='UserID')  # Field Name made lowercase.
+	OrderDelivered = models.BooleanField(db_column='OrderDelivered', default=False)  # Field Name made lowercase.
+	OrderTimestamp = models.DateTimeField(auto_now_add=True)
+	class Meta:
+		db_table = 'PastOrder'
 
 class OrderItems(models.Model):
-    OrderItemID = models.AutoField(db_column='OrderItemID', primary_key=True)  # Field Name made lowercase.
-    OrderID = models.ForeignKey(PastOrder, models.DO_NOTHING, db_column='OrderID', blank=True, null=True)  # Field Name made lowercase.
-    PriceID = models.ForeignKey('Price', models.DO_NOTHING, db_column='PriceID', blank=True, null=True)  # Field Name made lowercase.
-    Quantity = models.DecimalField(db_column='Quantity', max_digits=4, decimal_places=2, blank=True, null=True)  # Field Name made lowercase.
-    Notes = models.CharField(max_length=255, blank=True, null=True)
-    class Meta:
-        db_table = 'OrderItems'
+	OrderItemID = models.AutoField(db_column='OrderItemID', primary_key=True)  # Field Name made lowercase.
+	OrderID = models.ForeignKey(PastOrder, models.DO_NOTHING, db_column='OrderID', blank=True, null=True)  # Field Name made lowercase.
+	PriceID = models.ForeignKey('Price', models.DO_NOTHING, db_column='PriceID', blank=True, null=True)  # Field Name made lowercase.
+	Quantity = models.DecimalField(db_column='Quantity', max_digits=4, decimal_places=2, blank=True, null=True)  # Field Name made lowercase.
+	Notes = models.CharField(max_length=255, blank=True, null=True)
+	class Meta:
+		db_table = 'OrderItems'
 
 class Price(models.Model):
-    PriceID = models.AutoField(db_column='PriceID', primary_key=True)  # Field Name made lowercase.
-    ShopID = models.ForeignKey('Shop', models.DO_NOTHING, db_column='ShopID', blank=True, null=True)  # Field Name made lowercase.
-    ProductID = models.ForeignKey('Product', models.DO_NOTHING, db_column='ProductID', blank=True, null=True)  # Field Name made lowercase.
-    Price = models.DecimalField(db_column='Price', max_digits=8, decimal_places=2, blank=True, null=True)  # Field Name made lowercase.
+	PriceID = models.AutoField(db_column='PriceID', primary_key=True)  # Field Name made lowercase.
+	ShopID = models.ForeignKey('Shop', models.DO_NOTHING, db_column='ShopID', blank=True, null=True)  # Field Name made lowercase.
+	ProductID = models.ForeignKey('Product', models.DO_NOTHING, db_column='ProductID', blank=True, null=True)  # Field Name made lowercase.
+	Price = models.DecimalField(db_column='Price', max_digits=8, decimal_places=2, blank=True, null=True)  # Field Name made lowercase.
 
-    class Meta:
-        db_table = 'Price'
-        unique_together = ('ShopID', 'ProductID')
+	class Meta:
+		db_table = 'Price'
+		unique_together = ('ShopID', 'ProductID')
 
 class Product(models.Model):
-    ProductID = models.AutoField(db_column='ProductID', primary_key=True)  # Field Name made lowercase.
-    WordpressID = models.IntegerField(null=True, blank=True)
-    ProductTypeID = models.ForeignKey('ProductType', models.DO_NOTHING, db_column='ProductTypeID', blank=True, null=True)  # Field Name made lowercase.
-    ProductName = models.CharField(db_column='ProductName', max_length=45, blank=True, null=True)  # Field Name made lowercase.
-    ProductBrandID = models.ForeignKey(Brand, models.DO_NOTHING, db_column='ProductBrandID', blank=True, null=True)  # Field Name made lowercase.
-    ProductQuantity = models.FloatField(db_column='ProductQuantity', blank=True, null=True)  # Field Name made lowercase.
-    ProductUnit = models.CharField(db_column='ProductUnit', max_length=10, blank=True, null=True)  # Field Name made lowercase.
-    ProductBarcode = models.CharField(db_column='ProductBarcode', max_length=18, blank=True, null=True)  # Field Name made lowercase.
-    ProductWeight = models.DecimalField(null=False, default=1.0, decimal_places=2, max_digits=3)
-    class Meta:
-        db_table = 'Product'
+	ProductID = models.AutoField(db_column='ProductID', primary_key=True)  # Field Name made lowercase.
+	WordpressID = models.IntegerField(null=True, blank=True)
+	ProductTypeID = models.ForeignKey('ProductType', models.DO_NOTHING, db_column='ProductTypeID', blank=True, null=True)  # Field Name made lowercase.
+	ProductName = models.CharField(db_column='ProductName', max_length=45, blank=True, null=True)  # Field Name made lowercase.
+	ProductBrandID = models.ForeignKey(Brand, models.DO_NOTHING, db_column='ProductBrandID', blank=True, null=True)  # Field Name made lowercase.
+	ProductQuantity = models.FloatField(db_column='ProductQuantity', blank=True, null=True)  # Field Name made lowercase.
+	ProductUnit = models.CharField(db_column='ProductUnit', max_length=10, blank=True, null=True)  # Field Name made lowercase.
+	ProductBarcode = models.CharField(db_column='ProductBarcode', max_length=18, blank=True, null=True)  # Field Name made lowercase.
+	ProductWeight = models.FloatField(null=False, default=1.0)
+	class Meta:
+		db_table = 'Product'
 
 class ProductType(models.Model):
-    ProductTypeID = models.AutoField(db_column='ProductTypeID', primary_key=True)  # Field Name made lowercase.
-    ProductTypeName = models.CharField(db_column='ProductTypeName', max_length=45, unique=True)  # Field Name made lowercase.
-    CategoryID = models.ForeignKey(Category, models.DO_NOTHING, db_column='CategoryID', blank=True, null=True)  # Field Name made lowercase.
+	ProductTypeID = models.AutoField(db_column='ProductTypeID', primary_key=True)  # Field Name made lowercase.
+	ProductTypeName = models.CharField(db_column='ProductTypeName', max_length=45, unique=True)  # Field Name made lowercase.
+	CategoryID = models.ForeignKey(Category, models.DO_NOTHING, db_column='CategoryID', blank=True, null=True)  # Field Name made lowercase.
 
-    class Meta:
-        db_table = 'ProductType'
+	class Meta:
+		db_table = 'ProductType'
 
 class Shop(models.Model):
-    ShopID = models.AutoField(db_column='ShopID', primary_key=True)  # Field Name made lowercase.
-    ShopTypeID = models.ForeignKey('ShopType', models.DO_NOTHING, db_column='ShopTypeID')  # Field Name made lowercase.
-    ShopName = models.CharField(db_column='ShopName', max_length=45)  # Field Name made lowercase.
-    ShopLatitude = models.DecimalField(db_column='ShopLatitude', max_digits=10, decimal_places=8, blank=True, null=True)  # Field Name made lowercase.
-    ShopLongitude = models.DecimalField(db_column='ShopLongitude', max_digits=11, decimal_places=8, blank=True, null=True)  # Field Name made lowercase.
-    class Meta:
-        db_table = 'Shop'
+	ShopID = models.AutoField(db_column='ShopID', primary_key=True)  # Field Name made lowercase.
+	ShopTypeID = models.ForeignKey('ShopType', models.DO_NOTHING, db_column='ShopTypeID')  # Field Name made lowercase.
+	ShopName = models.CharField(db_column='ShopName', max_length=45)  # Field Name made lowercase.
+	ShopLatitude = models.DecimalField(db_column='ShopLatitude', max_digits=10, decimal_places=8, blank=True, null=True)  # Field Name made lowercase.
+	ShopLongitude = models.DecimalField(db_column='ShopLongitude', max_digits=11, decimal_places=8, blank=True, null=True)  # Field Name made lowercase.
+	class Meta:
+		db_table = 'Shop'
 
 class ShopType(models.Model):
-    ShopTypeID = models.AutoField(db_column='ShopTypeID', primary_key=True)  # Field Name made lowercase.
-    ShopTypeName = models.CharField(db_column='ShopTypeName', max_length=45)  # Field Name made lowercase.
+	ShopTypeID = models.AutoField(db_column='ShopTypeID', primary_key=True)  # Field Name made lowercase.
+	ShopTypeName = models.CharField(db_column='ShopTypeName', max_length=45)  # Field Name made lowercase.
 
-    class Meta:
-        db_table = 'ShopType'
+	class Meta:
+		db_table = 'ShopType'
 
 class ShoppingItem(models.Model):
-    UserID = models.ForeignKey('User', models.DO_NOTHING, db_column='UserID')  # Field Name made lowercase.
-    PriceID = models.ForeignKey(Price, models.DO_NOTHING, db_column='PriceID')  # Field Name made lowercase.
-    Quantity = models.DecimalField(db_column='Quantity', max_digits=4, decimal_places=2, blank=True, null=True)  # Field Name made lowercase.
-    ShoppingItemID = models.AutoField(db_column='ShoppingItemID', primary_key=True)  # Field Name made lowercase.
-    Notes = models.CharField(max_length=255, blank=True, null=True)
+	UserID = models.ForeignKey('User', models.DO_NOTHING, db_column='UserID')  # Field Name made lowercase.
+	PriceID = models.ForeignKey(Price, models.DO_NOTHING, db_column='PriceID')  # Field Name made lowercase.
+	Quantity = models.DecimalField(db_column='Quantity', max_digits=4, decimal_places=2, blank=True, null=True)  # Field Name made lowercase.
+	ShoppingItemID = models.AutoField(db_column='ShoppingItemID', primary_key=True)  # Field Name made lowercase.
+	Notes = models.CharField(max_length=255, blank=True, null=True)
 
-    class Meta:
-        db_table = 'ShoppingItem'
-        unique_together = ('UserID', 'PriceID',)
+	class Meta:
+		db_table = 'ShoppingItem'
+		unique_together = ('UserID', 'PriceID',)
 
 class User(models.Model):
-    UserID = models.AutoField(db_column='UserID', primary_key=True)  # Field Name made lowercase.
-    UserName = models.CharField(db_column='UserName', max_length=45, blank=True, null=True)  # Field Name made lowercase.
-    Userphonenumber = models.CharField(db_column='UserPhoneNumber', max_length=40, blank=True, null=False, unique=True)  # Field Name made lowercase.
-    Userlatitude = models.DecimalField(db_column='UserLatitude', max_digits=10, decimal_places=8, blank=True, null=True)  # Field Name made lowercase.
-    Userlongitude = models.DecimalField(db_column='UserLongitude', max_digits=11, decimal_places=8, blank=True, null=True)  # Field Name made lowercase.
-    UserState = models.IntegerField(default=99)
-    UserShopID = models.ForeignKey(Shop, models.DO_NOTHING, blank=True, null=True)
-    UserMessenger = models.BooleanField(default=False)
-    class Meta:
-        db_table = 'User'
+	UserID = models.AutoField(db_column='UserID', primary_key=True)  # Field Name made lowercase.
+	UserName = models.CharField(db_column='UserName', max_length=45, blank=True, null=True)  # Field Name made lowercase.
+	Userphonenumber = models.CharField(db_column='UserPhoneNumber', max_length=40, blank=True, null=False, unique=True)  # Field Name made lowercase.
+	Userlatitude = models.DecimalField(db_column='UserLatitude', max_digits=10, decimal_places=8, blank=True, null=True)  # Field Name made lowercase.
+	Userlongitude = models.DecimalField(db_column='UserLongitude', max_digits=11, decimal_places=8, blank=True, null=True)  # Field Name made lowercase.
+	UserState = models.IntegerField(default=99)
+	UserShopID = models.ForeignKey(Shop, models.DO_NOTHING, blank=True, null=True)
+	UserMessenger = models.BooleanField(default=False)
+	class Meta:
+		db_table = 'User'
 
 class Search(models.Model):
-    SearchID = models.AutoField(primary_key=True)
-    SearchTerm = models.CharField(max_length=50)
-    UserID = models.ForeignKey(User, models.CASCADE)
-    SearchTimestamp = models.DateTimeField(auto_now_add=True)
+	SearchID = models.AutoField(primary_key=True)
+	SearchTerm = models.CharField(max_length=50)
+	UserID = models.ForeignKey(User, models.CASCADE)
+	SearchTimestamp = models.DateTimeField(auto_now_add=True)
